Allow a `prompt` key in config files
The `prompt` key allows the user to specify a command with which to
populate the command line in the specified window or pane. The
`prompt` command is populated after all commands in the `command` array
are executed, and remains in the prompt field, awaiting manual
execution by the user.

This can be useful for autmated workflows which populate a user's
terminal window with several panes, each ready to display output from
a given command that should be run with special care/attention. For
example, a user may create a config file to create panes which tail log
files for a service, as well as a pane with a long/complex command
pre-populated, which would deploy changes to that service when executed.